// src/generated/models/Cgmp_notificationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_notificationscgmp_category = {
  100000000: 'ReviewRequest',
  100000001: 'UATReminder',
  100000002: 'Escalation',
  100000003: 'GIICCHandover',
  100000004: 'Closure',
  100000005: 'Emergency',
  100000006: 'System'
} as const;
export type Cgmp_notificationscgmp_category = keyof typeof Cgmp_notificationscgmp_category;
export const Cgmp_notificationscgmp_priority = {
  100000000: 'High',
  100000001: 'Medium',
  100000002: 'Low'
} as const;
export type Cgmp_notificationscgmp_priority = keyof typeof Cgmp_notificationscgmp_priority;
export const Cgmp_notificationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_notificationsstatecode = keyof typeof Cgmp_notificationsstatecode;
export const Cgmp_notificationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_notificationsstatuscode = keyof typeof Cgmp_notificationsstatuscode;

export interface Cgmp_notificationsBase {
  cgmp_acknowledgedat?: string;
  cgmp_acknowledgedby?: string;
  cgmp_actionurl?: string;
  cgmp_category: Cgmp_notificationscgmp_category;
  cgmp_isdismissed?: boolean;
  cgmp_isread?: boolean;
  cgmp_message?: string;
  cgmp_notificationid: string;
  cgmp_priority: Cgmp_notificationscgmp_priority;
  cgmp_recipientid?: string;
  cgmp_relatedchangeid?: string;
  cgmp_relatedprojectid?: string;
  cgmp_snoozeduntil?: string;
  cgmp_title: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_notificationsstatecode;
  statuscode?: Cgmp_notificationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_notifications extends Cgmp_notificationsBase {
  cgmp_categoryname?: string;
  cgmp_isdismissedname?: string;
  cgmp_isreadname?: string;
  cgmp_priorityname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}